Adds configuration option to mirror ForceAuthn
By default when the SATOSA SAML frontend receives a SAML authentication
request with ForceAuthn set to `True`, this information is not mirrored
in the SAML authentication request that is generated by the SATOSA SAML
backend towards the upstream identity provider. If the configuration
parameter `mirror_saml_forceauthn` is set to `True`, then the default
behaviour changes and the SATOSA SAML backend will set ForceAuthn to
true when it proxies a SAML authentication request with ForceAuthn set
to `True`. The default behaviour is `False`.